requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pdf2image>=1.16.0
tkinter-tooltip>=2.1.0
